chore: Replace look_ahead with new implementation
